Revamp library layout and add shelf selection dialog
