+++ src/app/pages/book-detail/book-detail.component.ts
@@ -2,25 +2,40 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { Book } from '../../shared/components/book';
 import { LibraryService } from '../../services/library.service';
+import { ShelfService } from '../../services/shelf.service';
 
 @Component({
   selector: 'app-book-detail',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, FormsModule],
   templateUrl: './book-detail.component.html',
   styleUrls: ['./book-detail.component.css']
 })
 export class BookDetailComponent implements OnInit {
   book: Book | undefined;
   apiUrl = 'http://localhost:3000/api/books';
-  constructor(private route: ActivatedRoute, private http: HttpClient, private library: LibraryService) {}
+  shelves: any[] = [];
+  selectedShelfId: number | null = null;
+  showDialog = false;
+
+  constructor(
+    private route: ActivatedRoute,
+    private http: HttpClient,
+    private library: LibraryService,
+    private shelfService: ShelfService
+  ) {}
 
   ngOnInit(): void {
     const googleBooksId = this.route.snapshot.paramMap.get('id');
     this.http.get<Book>(`${this.apiUrl}/google/${googleBooksId}`).subscribe(data => {
       this.book = data;
+    });
+    this.shelfService.getShelves().subscribe({
+      next: s => (this.shelves = s),
+      error: err => console.error('Failed to load shelves', err),
     });
   }
 
@@ -41,8 +56,11 @@ export class BookDetailComponent implements OnInit {
   }
 
   addToLibrary(): void {
-    if (!this.book) return;
+    this.showDialog = true;
+  }
 
+  confirmAdd(): void {
+    if (!this.book || !this.selectedShelfId) return;
     const payload = {
       googleBooksId: this.book.id,
       title: this.book.volumeInfo.title,
@@ -52,17 +70,25 @@ export class BookDetailComponent implements OnInit {
       pageCount: this.book.volumeInfo.pageCount,
       language: this.book.volumeInfo.language,
       thumbnailUrl: this.book.volumeInfo.imageLinks?.thumbnail,
-      previewLink: '', // Not provided by API response
+      previewLink: '',
       infoLink: this.book.volumeInfo.infoLink,
       averageRating: this.book.volumeInfo.averageRating,
       ratingsCount: this.book.volumeInfo.ratingsCount,
       isbn10: '',
-      isbn13: ''
+      isbn13: '',
+      authors: this.book.volumeInfo.authors,
+      categories: this.book.volumeInfo.categories,
     };
-
-    this.library.addBook(payload).subscribe({
-      next: () => alert('Book added to your library!'),
-      error: err => alert('Failed to add book: ' + err.message)
+    this.shelfService.addBookToShelf(this.selectedShelfId, payload).subscribe({
+      next: () => {
+        this.showDialog = false;
+        alert('Book added to shelf!');
+      },
+      error: err => alert('Failed to add book: ' + err.message),
     });
+  }
+
+  cancelAdd(): void {
+    this.showDialog = false;
   }
 }
