+++ src/app/pages/library/library.component.html
@@ -1,0 +1,38 @@
+<div class="library-page" *ngIf="username">
+  <h2>{{ username }}'s Library</h2>
+
+  <div class="controls">
+    <input type="text" [(ngModel)]="searchTerm" placeholder="Search books" />
+    <select [(ngModel)]="selectedCategory">
+      <option value="">All Categories</option>
+      <option *ngFor="let cat of categories" [value]="cat">{{ cat }}</option>
+    </select>
+  </div>
+
+  <section class="recent-section">
+    <h3>Recently Added</h3>
+    <div class="books-grid">
+      <app-book-card *ngFor="let b of filteredBooks() | slice:0:5" [book]="b"></app-book-card>
+    </div>
+  </section>
+
+  <section *ngIf="shelves.length" class="shelf-preview">
+    <div class="shelf-header">
+      <h3>Your Shelves</h3>
+      <a *ngIf="!showAllShelves" (click)="showAllShelves = true" class="see-all">See all shelves</a>
+    </div>
+    <div class="shelf-grid">
+      <div class="shelf" *ngFor="let s of shelves | slice:0:(showAllShelves ? shelves.length : 6)">
+        <h4>{{ s.name }}</h4>
+        <div class="shelf-books">
+          <app-book-card *ngFor="let sb of s.books | slice:0:3" [book]="sb"></app-book-card>
+        </div>
+      </div>
+    </div>
+  </section>
+
+  <div class="create-shelf">
+    <input type="text" [(ngModel)]="newShelfName" placeholder="New shelf name" />
+    <button (click)="createShelf()">Create Shelf</button>
+  </div>
+</div>
